reader tweak: reader macro . and :
As . and : cannot appear in regular identifiers I use them in
initializers, .bla is translated into (key 'id bla) and :bla is
translated into (key 'num bla).  In the end, they are used to produce
the following code:

    { [CONST-EXPR] = { .id = ... } }

The consequence of this change is that it is not possible to start a
name with a dot but this should not be a problem.  Moreover, this does
not alter the feature of writing bla[i].x as long as it can be written
without spaces.